Avoid computation of whole enummap in innermost loops
Essentially, we want all payout/payouthints/etc in two variants:
* One that computes the value for one specific EPlayerIndex.
* One that computes the values for all EPlayerIndex::values, returning
  them in an EnumMap.

We want this distinction because computing all values is expensive in
AI. (By computing only one value in payout/payouthints, we can reduce
execution time of determine_best_card (in tests) from over 11 to under
10 seconds.)

To achieve this distinction, we parametrize all relevant methods by a
TPerEPI. As of now, the only TPerEPI is EPlayerIndex, but something like
SAllEPI seems reasonable, too.

There are some technical limitations that will hit us once we actually
want to implement SAllEPI:
* Generalizing over single value/multiple values does not appear to be
  straightforward. We may be able to use GADTs one day, but until then,
  we would possibly need heavy workarounds.
* E.g. in TRules, we cannot use TPerEPI (if we do not want to resort to
  trait objects).

We try to keep the code as general as possible in this regard so that
future changes are easier to incorporate.