New subcommand to download played games
This is a very crude draft to implement a downloader that can be used to
collect game data suitable for e.g. AI training. Nonetheless, it feels
pretty fast and is probably good enough to be useful.

Note, however, that it does not follow best practices wrt user/password
handling.