should_stoss returns bool, not BOOL
I read the decompiled thing, and saw that it only returns 1 or 0, so bool should be justified.

However, is is a general problem to interpret return values from undocumented functions, and ideally I'd appreciate automated help when doing this. One idea might be that call_original asserts that the returned value "looks as expected".